Department class
The user should maintain the Department Rates data for each department via a menu
with the following options:
i. Add: This allows the user to add a new department rates record to the system.
ii. Update: This allows the user to update an existing department record.
iii. View: This allows the user to view a single department record.
iv. View All: This allows the user to view all department records.